requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
google-cloud-logging==3.11.3
google-cloud-firestore==2.19.0
PyJWT==2.8.0
python-dotenv==1.0.1
slowapi==0.1.9
bleach==6.1.0
pydantic==2.8.2
jinja2==3.1.4
python-multipart==0.0.9
httpx==0.27.2
pytest==8.2.0
pytest-asyncio==0.24.0
